Removed ISqlImportOperation
